% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_na_processing_formulas}
\alias{apply_na_processing_formulas}
\title{Apply NA Processing Formulas}
\usage{
apply_na_processing_formulas(df, list_formulas)
}
\arguments{
\item{df}{A data frame containing the data to which the formulas should be applied.}

\item{list_formulas}{A named list of formulas, with each formula represented as a string.
E.g. list(
X = "!is.na(a) ~ a, !is.na(b) ~ b, TRUE ~ c",
Y = "!is.na(m) ~ m, !is.na(l) ~ l, TRUE ~ NA"
)}
}
\value{
A data frame that is a copy of the input data frame, but with additional columns
for each formula in the input list. The name of each new column is the name of
the corresponding formula in the input list.
}
\description{
This function takes a data.frame and a list of strings containing formulas as input and applies the formulas to process the data.frame.
}
