% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{check_col_names_are_equal}
\alias{check_col_names_are_equal}
\title{Check that the names in two dataframes are equal}
\usage{
check_col_names_are_equal(df_estimates_microdata, na_table)
}
\description{
Check that the names in two dataframes (estimates from microdata and NA table) are equal,
excluding the \code{ccyy} and \code{ccyyd} columns.
}
