% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{compute_coverage_ratios}
\alias{compute_coverage_ratios}
\title{Compute National Accounts coverage ratios}
\usage{
compute_coverage_ratios(
  df_estimates_microdata,
  na_table,
  formulas_ratios =
    as.list(set_names(names(nationalaccountslis::lis_dashboard_microdata_formulas)))
)
}
\arguments{
\item{df_estimates_microdata}{A dataframe with the National Accounts estimates from the microdata.
Must contain the \code{ccyyd} column and one column per variable, with column names matching the
formulas in \code{formulas_microdata}. E.g. as returned by \code{compute_estimates_lis_microdata()}.}

\item{na_table}{A dataframe with the National Accounts data. E.g. as returned by \code{produce_oecd_na_table()}.
Must contain the following columns: \code{country}, \code{year} and one column per variable. E.g.
data.frame(country = c("ITA", "USA"), year = c(1985, 1985), D11R = c(100, 200), D12R = c(300, 400))}

\item{formulas_ratios}{A list with strings containing formulas to produce the National Accounts ratios.}
}
\value{
A dataframe with the National Accounts ratios.
}
\description{
Compute the National Accounts coverage ratios from the microdata estimates and items in National Accounts.
}
