% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates_microdata.R
\name{compute_estimates_lis_microdata}
\alias{compute_estimates_lis_microdata}
\title{Compute National Account estimates from LIS microdata}
\usage{
compute_estimates_lis_microdata(
  datasets,
  data_path,
  weights,
  formulas_microdata = nationalaccountslis::lis_dashboard_microdata_formulas
)
}
\arguments{
\item{datasets}{A vector of strings with the names of the datasets to read. The
fourth character must identify the database of the dataset (e.g. 'i' for LIS,
'w' for LWS, 'e' for ERFLIS). E.g. 'it14i'}

\item{data_path}{A string or vector of strings with the path to the folder containing the LIS datasets.
Must either have length 1 (all datasets in the same path) or the same length as datasets.}

\item{weights}{A string or vector of strings with the names of the weight variable(s) to use.
Must either have length 1 (all datasets use the same weight variable) or the same length as formulas.}

\item{formulas_microdata}{A list of formulas to compute the National Accounts estimates.}
}
\value{
A list of dataframes with the National Accounts estimates from the microdata.
}
\description{
Computes National Account estimates from LIS microdata. It grosses up variables
using population weights to reproduce the National Accounts numbers.
}
