% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates_microdata.R
\name{compute_gross_up_estimate}
\alias{compute_gross_up_estimate}
\title{Compute grossed up estimates from LIS microdata}
\usage{
compute_gross_up_estimate(df, weights, formulas)
}
\arguments{
\item{df}{A dataframe with the LIS microdata.}

\item{weights}{A string or vector of strings with the names of the weight variable(s) to use.
Must either have length 1 (all datasets use the same weight variable) or the same length as formulas.}

\item{formulas}{A list of strings witht the formulas to compute the National Accounts estimates.}
}
\description{
Compute grossed up estimates from LIS microdata
}
