% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{compute_national_accounts_ratios_oecd}
\alias{compute_national_accounts_ratios_oecd}
\title{Compute National Accounts ratios}
\usage{
compute_national_accounts_ratios_oecd(
  df_estimates_microdata,
  produce_oecd_na_table
)
}
\arguments{
\item{df_estimates_microdata}{A dataframe with the National Accounts estimates from the microdata.
E.g. as returned by \code{compute_estimates_lis_microdata()}.}

\item{na_data}{A dataframe with the National Accounts data. E.g. as returned by \code{download_na_oecd()}.
Must contain the following columns: \code{country}, \code{variable}, \code{sector}, \code{year}, \code{value}.}
}
\value{
A dataframe with the National Accounts ratios.
}
\description{
Compute the National Accounts ratios from the items in microdata and National Accounts.
}
