% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_iso2_to_iso3}
\alias{convert_iso2_to_iso3}
\title{Convert ISO2 country codes to ISOW3}
\usage{
convert_iso2_to_iso3(countries)
}
\arguments{
\item{countries}{A vector of ISO2 country codes.}
}
\value{
A vector of ISO3 country codes.
}
\description{
Convert ISO2 country codes to ISOW3
}
\examples{
convert_iso2_to_iso3(c("US", "DE", "JP"))
}
