% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_iso3_to_iso2}
\alias{convert_iso3_to_iso2}
\title{Convert ISO3 country codes to ISO2}
\usage{
convert_iso3_to_iso2(countries)
}
\arguments{
\item{countries}{A vector of ISO3 country codes.}
}
\value{
A vector of ISO2 country codes.
}
\description{
Convert ISO3 country codes to ISO2
}
\examples{
\dontrun{
convert_iso3_to_iso2(c("USA", "DEU", "JPN"))
}
}
