% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_na.r
\name{download_na_oecd}
\alias{download_na_oecd}
\title{Download National Accounts from OECD REST API}
\usage{
download_na_oecd(
  start_year = 1967,
  countries = nationalaccountstools::dashboard_countries,
  items = nationalaccountslis::dashboard_items
)
}
\arguments{
\item{start_year}{The first year to download data for.}

\item{countries}{A vector of ISO2 or ISO3 country codes. E.g.
c("AUS", "AUT", "BEL"). Defaults to the countries used in the LIS Compare.IT}

\item{items}{A vector of National Account items to download. E.g.
c("NFD11R", "NFD12R", "NFB3GR").
Defaults to the items used in the LIS Compare.IT Dashboard.}
}
\value{
A data frame with the following columns: country, variable,
sector, year, value.
}
\description{
Download National Accounts from OECD REST API.
}
\examples{
download_na_oecd(start_year = 1967, countries = c("AUS", "AUT", "BEL"),
 items = c("NFD11R", "NFD12R", "NFB3GR"))
}
