% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_all_vars_from_formulas}
\alias{parse_all_vars_from_formulas}
\title{Parse All Variables from Formulas}
\usage{
parse_all_vars_from_formulas(formulas)
}
\arguments{
\item{formulas}{A list of formulas, with each formula represented as a string.}
}
\value{
A character vector with all variables used in the formulas.
}
\description{
This function takes a list of formulas as input and returns all variables used in the formulas.
}
