% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{read_lis_dataset}
\alias{read_lis_dataset}
\alias{read_lws_dataset}
\title{Read a LIS dataset}
\usage{
read_lis_dataset(dataset, data_path, needed_vars)

read_lws_dataset(dataset, data_path, needed_vars)
}
\arguments{
\item{dataset}{A string with the name of the dataset to read. E.g. 'it14i'}

\item{data_path}{A string with the path to the folder containing the LIS datasets.}

\item{needed_vars}{A vector of strings with the names of the variables needed for the
National Accounts estimates.}
}
\value{
A dataframe with the variables needed for the National Accounts estimates.
}
\description{
Reads a LIS dataset and returns a dataframe with the variables needed for the
National Accounts estimates.
}
\section{Functions}{
\itemize{
\item \code{read_lws_dataset()}: Read a LWS dataset

}}
