% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{read_lis_microdata}
\alias{read_lis_microdata}
\title{Read a LIS or LWS dataset}
\usage{
read_lis_microdata(dataset, data_path, weights, formulas)
}
\arguments{
\item{dataset}{A string with the name of the dataset to read. E.g. 'it14i'. The
fifth character must identify the database of the dataset (e.g. 'i' for LIS,
'w' for LWS, 'e' for ERFLIS).}

\item{data_path}{A string with the path to the folder containing the LIS datasets.}

\item{formulas}{A list of formulas to compute the National Accounts estimates.}
}
\value{
A dataframe with the variables needed for the National Accounts estimates.
}
\description{
Reads a LIS or LWS dataset and returns a dataframe with the variables needed for the
National Accounts estimates.
}
