% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_formulas}
\alias{compute_formulas}
\title{Compute Formulas}
\usage{
compute_formulas(df, formulas)
}
\arguments{
\item{df}{A data frame containing the data to which the formulas should be applied.}

\item{formulas}{A named list of formulas, with each formula represented as a string.}
}
\value{
A data frame that is a copy of the input data frame, but with additional columns
for each formula in the input list. The name of each new column is the name of
the corresponding formula in the input list.
}
\description{
This function evaluates each formula in a list of formulas with respect to a data frame,
adding the results as new columns to the data frame.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9, d = 10:12)
formulas <- list(Z = "a + b", Y = "c+d")
compute_formulas(df, formulas)
}
